Move all items related to the shop page into a separate module.
