+++ src/ui/shop/mod.rs
@@ -1,0 +1,5 @@
+mod shop_page;
+mod shop_item;
+
+pub use shop_page::ShopPage;
+pub use shop_item::ShopItem;
